Сделала правки
