--- src/main/java/com/example/AsexualException.java
@@ -1,13 +1,0 @@
-package com.example;
-
-public class AsexualException extends Exception {
-    private String message;
-    public AsexualException(String message){
-        this.message = message;
-    };
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
