+++ src/main/java/com/example/Lion.java
@@ -1,8 +1,5 @@
 package com.example;
 
-import java.util.List;
-import com.example.AsexualException;
-
 public class Lion {
     private boolean hasMane;
     String sex;
@@ -13,30 +10,11 @@ public class Lion {
             hasMane = true;
         } else if ("Самка".equals(sex)) {
             hasMane = false;
-        } else if ("Бесполый".equals(sex)) {
-            throw new AsexualException("Это только для людей");
-        }
-
-        else
+        } else
         {
             throw new Exception("Используйте допустимые значения пола животного - самец или самка");
         }
     }
-    public Lion(String sex) throws Exception{
-        if ("Самец".equals(sex)) {
-            hasMane = true;
-        } else if ("Самка".equals(sex)) {
-            hasMane = false;
-        } else if ("Бесполый".equals(sex)) {
-            throw new AsexualException("Это только для людей");
-        }
-
-        else
-        {
-            throw new Exception("Используйте допустимые значения пола животного - самец или самка");
-        }
-    }
-
 
 
     public int getKittens() {
