+++ src/test/java/AnimalTest.java
@@ -3,7 +3,7 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 
 
-public class AnimalExceptionTest {
+public class AnimalTest {
 
     @Test
     public void animalExceptionTest(){
@@ -15,5 +15,11 @@ public class AnimalExceptionTest {
         catch(Exception e) {
             assertEquals("Неизвестный вид животного, используйте значение Травоядное или Хищник", e.getMessage());
         }
+    }
+
+    @Test
+    public void animalFamilyTest() {
+        Animal animal = new Animal();
+        assertEquals("Существует несколько семейств: заячьи, беличьи, мышиные, кошачьи, псовые, медвежьи, куньи", animal.getFamily());
     }
 }
